Reorganized layout based on PR #47 feedback
