+++ rir/src/analysis/State.h
