+++ rir/src/analysis/analysis.h
