+++ rir/src/analysis/framework.h
