+++ rir/src/optimizers/stupid_inline.h
@@ -2,9 +2,9 @@
 #define RIR_STUPID_INLINE_H
 
 #include "R/RList.h"
-#include "dataflow.h"
-#include "framework/analysis.h"
-#include "framework/dispatchers.h"
+#include "analyzers/dataflow.h"
+#include "analysis/analysis.h"
+#include "analysis/dispatchers.h"
 #include "interpreter/interp_context.h"
 #include "ir/CodeEditor.h"
 #include "ir/Compiler.h"
